+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.9.2 (10/3/2018 - )
 * httpClient: tweak request/response to make it more low level style
 * api: refined query param bean behavior, empty value will be treated as null, refer to core.framework.impl.web.bean.QueryParamMapper 
+* management: added /_sys/vm, /_sys/thread, /_sys/heap to provide more detailed diagnostic info (similar result of jcmd) 
 
 ### 6.9.1 (10/3/2018)
 * httpClient: fix bug on shouldRetry